+++ source/udt/cpp/buffer.cpp
@@ -185,7 +185,7 @@ int CSndBuffer::addBufferFromFile(fstream& ifs, int len)
          pktlen = m_iMSS;
 
       ifs.read(s->m_pcData, pktlen);
-      if ((pktlen = ifs.gcount()) <= 0)
+      if ((pktlen = (int)ifs.gcount()) <= 0)
          break;
 
       // currently file transfer is only available in streaming mode, message is always in order, ttl = infinite
